Move all source files to root directory
