+++ inheritance.go
